Unify fw & johnson benchmarks code
